# swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getDefault","parameters":[],"responses":{"200":{"description":""}}}},"/espacio-academico":{"post":{"operationId":"EspacioAcademicoController_post","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Espacio_academicoDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"EspacioAcademicoController_getAll","parameters":[{"name":"query","required":false,"in":"query","description":"query - Filter. e.g. col1:v1,col2:v2 …","schema":{"type":"string"}},{"name":"fields","required":false,"in":"query","description":"fields - Fields returned. e.g. col1,col2 …","schema":{"type":"string"}},{"name":"sortby","required":false,"in":"query","description":"sortby - Sorted-by fields. e.g. col1,col2 …","schema":{"type":"string"}},{"name":"order","required":false,"in":"query","description":"order - Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc …","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","description":"limit - Limit the size of result set. Must be an integer","schema":{"type":"string"}},{"name":"offset","required":false,"in":"query","description":"offset - Start position of result set. Must be an integer","schema":{"type":"string"}},{"name":"populate","required":false,"in":"query","description":"populate - show subqueries. Must be a boolean","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/espacio-academico/{id}":{"get":{"operationId":"EspacioAcademicoController_getById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"put":{"operationId":"EspacioAcademicoController_put","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Espacio_academicoDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"EspacioAcademicoController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/espacio-academico-estudiantes":{"post":{"operationId":"EspacioAcademicoEstudiantesController_post","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Espacio_academico_estudiantesDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"EspacioAcademicoEstudiantesController_getAll","parameters":[{"name":"query","required":false,"in":"query","description":"query - Filter. e.g. col1:v1,col2:v2 …","schema":{"type":"string"}},{"name":"fields","required":false,"in":"query","description":"fields - Fields returned. e.g. col1,col2 …","schema":{"type":"string"}},{"name":"sortby","required":false,"in":"query","description":"sortby - Sorted-by fields. e.g. col1,col2 …","schema":{"type":"string"}},{"name":"order","required":false,"in":"query","description":"order - Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc …","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","description":"limit - Limit the size of result set. Must be an integer","schema":{"type":"string"}},{"name":"offset","required":false,"in":"query","description":"offset - Start position of result set. Must be an integer","schema":{"type":"string"}},{"name":"populate","required":false,"in":"query","description":"populate - show subqueries. Must be a boolean","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/espacio-academico-estudiantes/{id}":{"get":{"operationId":"EspacioAcademicoEstudiantesController_getById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"put":{"operationId":"EspacioAcademicoEstudiantesController_put","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Espacio_academico_estudiantesDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"EspacioAcademicoEstudiantesController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}}},"info":{"title":"espacios_academicos_crud","description":"API CRUD para la gestion de espacios académicos en el registro de notas","version":"1.0","contact":{}},"tags":[{"name":"espacio_academico","description":""}],"servers":[],"components":{"schemas":{"Espacio_academicoDto":{"type":"object","properties":{"nombre":{"type":"string"},"descripcion":{"type":"string"},"codigo_abreviacion":{"type":"string"},"codigo":{"type":"string"},"plan_estudio_id":{"type":"string"},"creditos":{"type":"number"},"distribucion_horas":{"type":"object"},"clasificacion_espacio_id":{"type":"number"},"espacios_requeridos":{"type":"object"},"grupo":{"type":"string"},"inscritos":{"type":"number"},"docente_id":{"type":"number"},"horario_id":{"type":"string"},"espacio_academico_padre":{"type":"string"},"activo":{"type":"boolean"},"fecha_creacion":{"format":"date-time","type":"string"},"fecha_modificacion":{"format":"date-time","type":"string"}},"required":["nombre","descripcion","codigo_abreviacion","codigo","plan_estudio_id","creditos","distribucion_horas","clasificacion_espacio_id","espacios_requeridos","grupo","inscritos","docente_id","horario_id","espacio_academico_padre","activo","fecha_creacion","fecha_modificacion"]},"Espacio_academico_estudiantesDto":{"type":"object","properties":{"espacio_academico_id":{"type":"string"},"estudiante_id":{"type":"number"},"periodo_id":{"type":"number"},"activo":{"type":"boolean"},"fecha_creacion":{"format":"date-time","type":"string"},"fecha_modificacion":{"format":"date-time","type":"string"}},"required":["espacio_academico_id","estudiante_id","periodo_id","activo","fecha_creacion","fecha_modificacion"]}}}}